feat(supervisor): adopt the new chat as the Sessions tab
Supervisor's Sessions tab now IS the chat surface — a findable, cross-workspace
list of your chat sessions (continue any from any device) + 'New chat with <agent>',
each opening the live ChatPanel. Replaces the emdash OpenSessions tail with the
bigger/better thing; the quick-dispatch Composer stays below.

- ChatSessionsPanel: reusable, cross-workspace (lists via the flat mount = all your
  workspaces; each session links to ITS OWN workspace's chat route; new chat is
  created in the chosen agent's workspace). Used by both supervisor and the chat home.
- createSession gains a workspace option (tenant route) for cross-workspace new-chat.
- ChatListPage is now a thin wrapper over the panel (DRY; fixes its cross-ws nav).
